* develop:
  fix update Question list
  muitple update sub - wip
  Question list Insert -Done
  ui - create sub
  fix table question
  option upload + link
  từ vựng - edit
  fixbug
  fixbug validate
  fixbug
  add toast noti
  fix type input search

# Conflicts:
#	config/common.php
#	resources/views/admin/layouts/script.blade.php

+++ app/Http/Controllers/Backend/VocabularyCatController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Backend;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\VocabularyCatRequest;
+use App\Http\Requests\VocabularyCatRequestUpdate;
 use App\Models\VocabularyCat;
 use App\Repositories\VocabularyCatRepository;
 use Illuminate\Http\Request;
@@ -19,7 +20,7 @@ class VocabularyCatController extends Controller
 
     public function index()
     {
-        $data = VocabularyCat::orderBy('id', 'DESC')->paginate(10);
+        $data = VocabularyCat::orderBy('id', 'DESC')->paginate(PAGE_SIZE);
         return view('admin.vocabularyCat.index', [
             'data' => $data,
         ]);
@@ -27,10 +28,10 @@ class VocabularyCatController extends Controller
 
     public function search(Request $request)
     {
-        $question = VocabularyCat::query();
+        $mSearch = VocabularyCat::query();
         if (!empty(request('keyword'))) {
-            $question->where('name', 'LIKE', '%' . request('keyword') . '%');
-            $question->orWhere('id', request('keyword'));
+            $mSearch->where('name', 'LIKE', '%' . request('keyword') . '%');
+            $mSearch->orWhere('id', request('keyword'));
         }
         if (!empty(request('rangeDate'))) {
             $temp = explode('-', request('rangeDate'));
@@ -41,13 +42,13 @@ class VocabularyCatController extends Controller
             $endDate = \Carbon\Carbon::createFromFormat('d/m/Y', $endDate)
                 ->format('Y-m-d 23:59:59');
 
-            $question->where('created_at', '>=', $startDate)
+            $mSearch->where('created_at', '>=', $startDate)
                 ->where('created_at', '<=', $endDate);
         }
         if (request('status') >= 0) {
-            $question->where('status', request('status'));
+            $mSearch->where('status', request('status'));
         }
-        $data = $question->orderBy('id', 'DESC')->paginate(10);
+        $data = $mSearch->orderBy('id', 'DESC')->paginate(10);
 
         return view('admin.vocabularyCat.index', [
             'data' => $data
@@ -63,11 +64,13 @@ class VocabularyCatController extends Controller
     {
         $data = $request->except('_token', 'files');
         if ($request->hasFile('thumb')) {
-            $path = $request->file('thumb')->store('thumbnails', 'public');
-            $data['thumb'] = $path;
+            if (!isUrl($request->thumb)) {
+                $path = $request->file('thumb')->store('thumbnails', 'public');
+                $data['thumb'] = $path;
+            }
         }
-        $this->vocabularyCatRepository->storeNew($data);
-        return redirect()->route('admin.vocabularyCat.index')->with('success', 'Thêm mới thành công');
+        $isSave = $this->vocabularyCatRepository->storeNew($data);
+        return redirect()->route('admin.vocabularyCat.index')->with($isSave ? SUCCESS : ERROR, $isSave ? CREATE_SUCCESS : CREATE_ERROR);
     }
 
     public function edit($id)
@@ -78,22 +81,25 @@ class VocabularyCatController extends Controller
         ]);
     }
 
-    public function update(Request $request)
+    public function update(VocabularyCatRequestUpdate $request)
     {
+        $request->request->remove('inlineRadioUpload');
         $detail = VocabularyCat::find($request->id);
         $data = $request->except('_token', 'files');
         if ($request->hasFile('thumb')) {
-            $path = $request->file('thumb')->store('thumbnails', 'public');
-            $data['thumb'] = $path;
             // remove old image
-            if (!empty($detail->thumb)) {
+            if (!empty($detail->thumb) && !isUrl($request->thumb)) {
                 if (file_exists('storage/' . $detail->thumb)) {
                     unlink('storage/' . $detail->thumb);
                 };
             }
+            if (!isUrl($request->thumb)) {
+                $path = $request->file('thumb')->store('thumbnails', 'public');
+                $data['thumb'] = $path;
+            }
         }
-        $this->vocabularyCatRepository->update($request->id, $data);
-        return redirect()->route('admin.vocabularyCat.index');
+        $isSave = $this->vocabularyCatRepository->update($request->id, $data);
+        return redirect()->route('admin.vocabularyCat.index')->with($isSave ? SUCCESS : ERROR, $isSave ? UPDATE_SUCCESS : UPDATE_ERROR);
     }
 
     public function remove(Request $request)
